Move config-related functions to separate interface
