+++ src/main/java/com/greghaskins/spectrum/Spectrum.java
@@ -4,10 +4,6 @@ import com.greghaskins.spectrum.dsl.specification.Specification;
 import com.greghaskins.spectrum.internal.DeclarationState;
 import com.greghaskins.spectrum.internal.Suite;
 import com.greghaskins.spectrum.internal.blocks.ConstructorBlock;
-import com.greghaskins.spectrum.internal.configuration.BlockFocused;
-import com.greghaskins.spectrum.internal.configuration.BlockIgnore;
-import com.greghaskins.spectrum.internal.configuration.BlockTagging;
-import com.greghaskins.spectrum.internal.configuration.ConfiguredBlock;
 import com.greghaskins.spectrum.internal.junit.Rules;
 
 import org.junit.runner.Description;
@@ -152,10 +148,6 @@ public final class Spectrum extends Runner {
     Specification.xit(behavior, block);
   }
 
-  public static Configuration configure() {
-    return new Configuration(DeclarationState.instance().getCurrentSuiteBeingDeclared());
-  }
-
 
   /**
    * Uses the given class as a mix-in for JUnit rules to be applied. These rules will cascade down
@@ -245,90 +237,6 @@ public final class Spectrum extends Runner {
    */
   public static <T> Supplier<T> let(final com.greghaskins.spectrum.ThrowingSupplier<T> supplier) {
     return Specification.let(supplier);
-  }
-
-  /**
-   * Surround a {@link com.greghaskins.spectrum.Block} with the {@code with} statement to add
-   * preconditions and metadata to it. E.g. <code>with(tags("foo"), () -&gt; {})</code>.<br>
-   * Note: preconditions and metadata can be chained using the
-   * {@link BlockConfigurationChain#and(BlockConfigurationChain)} method. E.g.
-   * <code>with(tags("foo").and(ignore()), () -&gt; {})</code>
-   *
-   * @param configuration the chainable block configuration
-   * @param block the enclosed block
-   * @return a wrapped block with the given configuration
-   *
-   * @see #ignore(String)
-   * @see #ignore()
-   * @see #focus()
-   * @see #tags(String...)
-   *
-   */
-  public static com.greghaskins.spectrum.Block with(final BlockConfigurationChain configuration,
-      final com.greghaskins.spectrum.Block block) {
-
-    return ConfiguredBlock.with(configuration.getBlockConfiguration(), block);
-  }
-
-  /**
-   * Mark a block as ignored by surrounding it with the ignore method.
-   *
-   * @param why explanation of why this block is being ignored
-   * @param block the block to ignore
-   * @return a wrapped block which will be ignored
-   */
-  public static com.greghaskins.spectrum.Block ignore(final String why,
-      final com.greghaskins.spectrum.Block block) {
-    return with(ignore(why), block);
-  }
-
-  /**
-   * Mark a block as ignored by surrounding it with the ignore method.
-   *
-   * @param block the block to ignore
-   * @return a wrapped block which will be ignored
-   */
-  public static com.greghaskins.spectrum.Block ignore(final com.greghaskins.spectrum.Block block) {
-    return with(ignore(), block);
-  }
-
-  /**
-   * Ignore the suite or spec.
-   *
-   * @return a chainable configuration that will ignore the block within a {@link #with}
-   */
-  public static BlockConfigurationChain ignore() {
-    return new BlockConfigurationChain().with(new BlockIgnore());
-  }
-
-  /**
-   * Ignore the suite or spec.
-   *
-   * @param reason why this block is ignored
-   * @return a chainable configuration that will ignore the block within a {@link #with}
-   */
-  public static BlockConfigurationChain ignore(final String reason) {
-    return new BlockConfigurationChain().with(new BlockIgnore(reason));
-  }
-
-  /**
-   * Tags the suite or spec that is being declared with the given strings. Depending on the current
-   * filter criteria, this may lead to the item being ignored during test execution.
-   *
-   * @param tags tags that relate to the suite or spec
-   * @return a chainable configuration that has these tags set for the block in {@link #with}
-   */
-  public static BlockConfigurationChain tags(final String... tags) {
-    return new BlockConfigurationChain().with(new BlockTagging(tags));
-  }
-
-  /**
-   * Marks the suite or spec to be focused.
-   *
-   * @return a chainable configuration that will focus the suite or spec in the {@link #with}
-   */
-  public static BlockConfigurationChain focus() {
-    return new BlockConfigurationChain().with(new BlockFocused());
   }
 
   private final Suite rootSuite;
